Trekkjer ut eigne klassar for initialisering av featuretoggle-opplegget, det gjer det enklare å jobbe med og det gjer at FeatureToggleConfig-klassa blir ryddigare og lettare å få oversikt over. Det er der vi gjer flest funksjonsbrytar-relaterte endringar, så det er fint å optimalisere for at endringar der blir enkelt. Gjenstår å få ut FeatureToggleService-interfacet også, men det får bli i ny commit
