+++ src/main/kotlin/no/nav/familie/ba/sak/config/featureToggle/FeatureToggleProperties.kt
@@ -1,0 +1,20 @@
+package no.nav.familie.ba.sak.config.featureToggle
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+import org.springframework.boot.context.properties.ConstructorBinding
+import java.net.URI
+
+@ConfigurationProperties("funksjonsbrytere")
+@ConstructorBinding
+class FeatureToggleProperties(
+    val enabled: Boolean,
+    val unleash: Unleash
+) {
+
+    @ConstructorBinding
+    data class Unleash(
+        val uri: URI,
+        val cluster: String,
+        val applicationName: String
+    )
+}
